+++ R/jam-igraph-internals.R
@@ -248,7 +248,7 @@ default_igraph_values <- function
       label.font=1,
       label.cex=1,
       frame.color="black",
-      frame.lwd=1,
+      # frame.lwd=1,
       frame.width=1,
       shape="circle",
       pie=1,